react!: move AssertLogger utility class to org.usefultoys.slf4jtestmock
