+++ src/main/java/org/slf4j/impl/MockLogger.java
@@ -21,6 +21,7 @@ import lombok.Setter;
 import org.slf4j.Logger;
 import org.slf4j.Marker;
 import org.slf4j.impl.MockLoggerEvent.Level;
+import org.usefultoys.slf4jtestmock.AssertLogger;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
