+++ src/main/java/org/usefultoys/slf4jtestmock/AssertLogger.java
@@ -13,12 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.slf4j.impl;
+package org.usefultoys.slf4jtestmock;
 
 import lombok.experimental.UtilityClass;
 import org.junit.jupiter.api.Assertions;
 import org.slf4j.Logger;
 import org.slf4j.Marker;
+import org.slf4j.impl.MockLogger;
+import org.slf4j.impl.MockLoggerEvent;
 import org.slf4j.impl.MockLoggerEvent.Level;
 
 import java.util.List;
